Move CDC code into 32blit lib for use in user apps.
